+++ trainers/coop.py
@@ -263,14 +263,14 @@ class CoOp(TrainerX):
         if prec == "amp":
             with autocast():
                 output = self.model(image)
-                loss = F.cross_entropy(output, label)
+                loss = F.cross_entropy(output.float(), label.long())
             self.optim.zero_grad()
             self.scaler.scale(loss).backward()
             self.scaler.step(self.optim)
             self.scaler.update()
         else:
             output = self.model(image)
-            loss = F.cross_entropy(output, label)
+            loss = F.cross_entropy(output.float(), label.long())
             self.model_backward_and_update(loss)
 
         loss_summary = {
